backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy>=2.0.39
alembic==1.13.2
pydantic>=2.10,<3
pydantic-settings>=2.6,<3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.2
